add confirm ti delete and also make some  styles  to comment and  the articles  page

+++ src/Components/Comments/CommentItem.tsx
@@ -25,8 +25,10 @@ const CommentItem = ({ comment, userId }: CommentForArticle) => {
 
     const handleDeleteComment = async () => {
         try {
-            await axios.delete(`${DOMAIN}/api/comments/${comment.id}`);
-            router.refresh();
+            if (confirm("Are You Want To Delete This Comment")) {
+                await axios.delete(`${DOMAIN}/api/comments/${comment.id}`);
+                router.refresh();
+            }
 
         }
         catch (error: any) {
@@ -41,11 +43,11 @@ const CommentItem = ({ comment, userId }: CommentForArticle) => {
         }}
             className="  flex flex-col w-full bg-white rounded-lg mb-4 p-3">
 
-            <div className="flex items-center justify-between mb-1">
-                <div className="flex justify-between items-center gap-2">
+            <div className="flex items-center justify-between mb-1 p-1 bg-gray-800 rounded-md text-white">
+                <div className="flex justify-between items-center gap-1">
                     <FaRegCircleUser size={30} />
 
-                    <h2 className="text-lg ml-2 font-semibold  text-gray-800 flex items-center ">
+                    <h2 className="text-lg ml-2 font-semibold  text-white flex items-center ">
 
 
                         {comment.user.name}
